refactor(pipeline): reorganize detection into a package structure
- Move detection modules into src/pipeline/detection/ directory.
- Rename files for clarity:
  - detection.py -> orchestrator.py
  - detection_hybrid.py -> hybrid.py
  - detection_config.py -> config.py
  - detection_utils.py -> utils.py
- Add __init__.py to expose public API from src.pipeline.detection.
- Update internal imports to use relative paths.
- Update external references in main.py and unit tests.
- Functional logic remains unchanged, verified with unit and smoke tests.
