+++ src/pipeline/detection/__init__.py
@@ -1,0 +1,11 @@
+"""Detection package providing barlines detection orchestration."""
+
+from .orchestrator import DetectorOrchestrator, run_detection_step
+from .utils import resolve_barlines_and_masks_config, resolve_paths_from_detection
+
+__all__ = [
+    "DetectorOrchestrator",
+    "run_detection_step",
+    "resolve_barlines_and_masks_config",
+    "resolve_paths_from_detection",
+]
