+++ src/pipeline/detection/config.py
